Add logic for RAM64, using RAM8.
64 and 8 refer to number of registers. Not word size. They are both same
word size = 16.

here also dmux acts something to select appropriate RAM
a good note here is that address like 3,4,5 are used to select RAM

while 0,1,2 goes within RAM.

Mux again selects one of the outputs.

Once again the mux and dmux share the address lines (which in this case
are 3,4,5)